Better formatting on difference
