+++ crates/store/re_format_arrow/src/lib.rs
@@ -4,13 +4,13 @@ use std::fmt::Formatter;
 
 use arrow::{
     array::{Array, ArrayRef, ListArray},
-    datatypes::{DataType, Field, Fields, IntervalUnit, TimeUnit},
+    datatypes::{DataType, Field, Fields},
     util::display::{ArrayFormatter, FormatOptions},
 };
 use comfy_table::{presets, Cell, Row, Table};
 use itertools::{Either, Itertools as _};
 
-use re_arrow_util::ArrowArrayDowncastRef as _;
+use re_arrow_util::{format_data_type, ArrowArrayDowncastRef as _};
 use re_tuid::Tuid;
 use re_types_core::Loggable as _;
 
@@ -75,122 +75,6 @@ fn parse_tuid(array: &dyn Array, index: usize) -> Option<Tuid> {
 }
 
 // ---
-
-// arrow has `ToString` implemented, but it is way too verbose.
-#[repr(transparent)]
-struct DisplayTimeUnit(TimeUnit);
-
-impl std::fmt::Display for DisplayTimeUnit {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let s = match self.0 {
-            TimeUnit::Second => "s",
-            TimeUnit::Millisecond => "ms",
-            TimeUnit::Microsecond => "us",
-            TimeUnit::Nanosecond => "ns",
-        };
-        f.write_str(s)
-    }
-}
-
-// arrow has `ToString` implemented, but it is way too verbose.
-#[repr(transparent)]
-struct DisplayIntervalUnit(IntervalUnit);
-
-impl std::fmt::Display for DisplayIntervalUnit {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let s = match self.0 {
-            IntervalUnit::YearMonth => "year/month",
-            IntervalUnit::DayTime => "day/time",
-            IntervalUnit::MonthDayNano => "month/day/nano",
-        };
-        f.write_str(s)
-    }
-}
-
-// arrow has `ToString` implemented, but it is way too verbose.
-#[repr(transparent)]
-struct DisplayDatatype<'a>(&'a DataType);
-
-impl std::fmt::Display for DisplayDatatype<'_> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let s = match &self.0 {
-            DataType::Null => "null",
-            DataType::Boolean => "bool",
-            DataType::Int8 => "i8",
-            DataType::Int16 => "i16",
-            DataType::Int32 => "i32",
-            DataType::Int64 => "i64",
-            DataType::UInt8 => "u8",
-            DataType::UInt16 => "u16",
-            DataType::UInt32 => "u32",
-            DataType::UInt64 => "u64",
-            DataType::Float16 => "f16",
-            DataType::Float32 => "f32",
-            DataType::Float64 => "f64",
-            DataType::Timestamp(unit, timezone) => {
-                let s = if let Some(tz) = timezone {
-                    format!("Timestamp({}, {tz})", DisplayTimeUnit(*unit))
-                } else {
-                    format!("Timestamp({})", DisplayTimeUnit(*unit))
-                };
-                return f.write_str(&s);
-            }
-            DataType::Date32 => "Date32",
-            DataType::Date64 => "Date64",
-            DataType::Time32(unit) => {
-                let s = format!("Time32({})", DisplayTimeUnit(*unit));
-                return f.write_str(&s);
-            }
-            DataType::Time64(unit) => {
-                let s = format!("Time64({})", DisplayTimeUnit(*unit));
-                return f.write_str(&s);
-            }
-            DataType::Duration(unit) => {
-                let s = format!("Duration({})", DisplayTimeUnit(*unit));
-                return f.write_str(&s);
-            }
-            DataType::Interval(unit) => {
-                let s = format!("Interval({})", DisplayIntervalUnit(*unit));
-                return f.write_str(&s);
-            }
-            DataType::Binary => "Binary",
-            DataType::FixedSizeBinary(size) => return write!(f, "FixedSizeBinary[{size}]"),
-            DataType::LargeBinary => "LargeBinary",
-            DataType::Utf8 => "Utf8",
-            DataType::LargeUtf8 => "LargeUtf8",
-            DataType::List(ref field) => {
-                let s = format!("List[{}]", Self(field.data_type()));
-                return f.write_str(&s);
-            }
-            DataType::FixedSizeList(field, len) => {
-                let s = format!("FixedSizeList[{}; {len}]", Self(field.data_type()));
-                return f.write_str(&s);
-            }
-            DataType::LargeList(field) => {
-                let s = format!("LargeList[{}]", Self(field.data_type()));
-                return f.write_str(&s);
-            }
-            DataType::Struct(fields) => return write!(f, "Struct[{}]", fields.len()),
-            DataType::Union(fields, _) => return write!(f, "Union[{}]", fields.len()),
-            DataType::Map(field, _) => return write!(f, "Map[{}]", Self(field.data_type())),
-            DataType::Dictionary(key, value) => {
-                return write!(f, "Dictionary{{{}: {}}}", Self(key), Self(value));
-            }
-            DataType::Decimal128(_, _) => "Decimal128",
-            DataType::Decimal256(_, _) => "Decimal256",
-            DataType::BinaryView => "BinaryView",
-            DataType::Utf8View => "Utf8View",
-            DataType::ListView(field) => return write!(f, "ListView[{}]", Self(field.data_type())),
-            DataType::LargeListView(field) => {
-                return write!(f, "LargeListView[{}]", Self(field.data_type()));
-            }
-            DataType::RunEndEncoded(_run_ends, values) => {
-                return write!(f, "RunEndEncoded[{}]", Self(values.data_type()));
-            }
-        };
-        f.write_str(s)
-    }
-}
 
 struct DisplayMetadata {
     prefix: &'static str,
@@ -422,13 +306,13 @@ fn format_dataframe_without_metadata(
                     Cell::new(format!(
                         "{}\n---\ntype: {}",
                         trim_name(field.name()),
-                        DisplayDatatype(field.data_type()),
+                        format_data_type(field.data_type()),
                     ))
                 } else {
                     Cell::new(format!(
                         "{}\n---\ntype: {}\n{}",
                         trim_name(field.name()),
-                        DisplayDatatype(field.data_type()),
+                        format_data_type(field.data_type()),
                         DisplayMetadata {
                             prefix: "",
                             metadata: field.metadata().clone().into_iter().collect()
