+++ crates/utils/re_arrow_util/src/lib.rs
@@ -3,7 +3,9 @@
 mod arrays;
 mod batches;
 mod compare;
+mod format_data_type;
 
 pub use self::arrays::*;
 pub use self::batches::*;
 pub use self::compare::*;
+pub use self::format_data_type::*;
